Replica Set 상세 조회 API 추가 및 공통 리소스 조회 API 변수형 수정

+++ src/main/java/kr/co/strato/adapter/k8s/common/proxy/InNamespaceProxy.java
@@ -1,10 +1,16 @@
 package kr.co.strato.adapter.k8s.common.proxy;
 
-import kr.co.strato.adapter.k8s.common.model.ClusterResourceInfo;
+import org.springframework.cloud.openfeign.FeignClient;
+import org.springframework.web.bind.annotation.DeleteMapping;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestParam;
+import org.springframework.web.bind.annotation.ResponseBody;
+
 import kr.co.strato.adapter.k8s.common.model.ResourceListSearchInfo;
 import kr.co.strato.adapter.k8s.common.model.WorkloadResourceInfo;
-import org.springframework.cloud.openfeign.FeignClient;
-import org.springframework.web.bind.annotation.*;
 
 @FeignClient(value="InNamespace", url = "${service.kubernetes-interface.url}")
 public interface InNamespaceProxy {
@@ -18,7 +24,7 @@ public interface InNamespaceProxy {
     @GetMapping("/inNamespace/{resourceType}")
     public @ResponseBody String getResource(
             @PathVariable("resourceType") String resourceType,
-            @RequestParam("kubeConfigId") Integer kubeConfigId,
+            @RequestParam("kubeConfigId") Long kubeConfigId,
             @RequestParam("namespace") String namespace,
             @RequestParam("name") String name);
